refactor: checkContentDuplicateAndPlusCategoryCount method 리팩토링 - #149
checkContentDuplicateAndAddCategorySaveLog으로 rename
custom repository로 옮기고 코드를 정리함
