+++ user_logs/1.txt
@@ -3,3 +3,5 @@
 {"categoryId": 16,"savedAt": 1674228317271}
 {"categoryId": 16,"savedAt": 1674646370377}
 {"categoryId": 14,"savedAt": 1674646406408}
+{"categoryId": 16,"savedAt": 1674647076221}
+{"categoryId": 14,"savedAt": 1674647103041}
